# secp256k1: Optimize normalize and NAF, correct normalize, and add tests. 