@jrick 2017-06-09
I'm unqualified to check the correctness of the code but I will do some testing downloading the mainnet and testnet chains without checkpoints.