@jrick 2017-06-09
```
19:26:13.437 2017-06-09 [INF] BMGR: Processed 1 block in the last 2m33.59s (5 transactions, 25 stake transactions, height 140684, 2017-06-09 19:26:08 +0000 UTC)
```

Mainnet synced without issues.