@jrick 2017-06-09
Testnet 2 synced as well